Updated testing + SpecAugment numerical issues
